refactor (registration-api-client) Rename RegistrationApiClient to RegistrationApiBindings
