+++ applications/api-cat/pom.xml
@@ -118,8 +118,8 @@
         </dependency>
         <dependency>
             <groupId>no.dcat</groupId>
-            <artifactId>registration-api-client</artifactId>
-            <version>1.6.0</version>
+            <artifactId>registration-api-bindings</artifactId>
+            <version>1.0.0</version>
         </dependency>
         <dependency>
             <groupId>no.dcat</groupId>
