+++ applications/api-cat/src/main/java/no/acat/service/RegistrationApiClient.java
@@ -1,12 +1,13 @@
 package no.acat.service;
 
+import no.fdk.registration.bindings.RegistrationApiBindings;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Service;
 
 
 @Service
-public class RegistrationApiClient extends no.dcat.client.registrationapi.RegistrationApiClient {
+public class RegistrationApiClient extends RegistrationApiBindings {
     @Autowired
     private Environment env;
