+++ applications/informationmodel-cat/pom.xml
@@ -94,8 +94,8 @@
         </dependency>
         <dependency>
             <groupId>no.dcat</groupId>
-            <artifactId>registration-api-client</artifactId>
-            <version>1.6.0</version>
+            <artifactId>registration-api-bindings</artifactId>
+            <version>1.0.0</version>
         </dependency>
         <dependency>
             <groupId>no.dcat</groupId>
