+++ applications/informationmodel-cat/src/main/java/no/fdk/imcat/service/RegistrationApiClient.java
@@ -1,12 +1,13 @@
 package no.fdk.imcat.service;
 
+import no.fdk.registration.bindings.RegistrationApiBindings;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Service;
 
 
 @Service
-public class RegistrationApiClient extends no.dcat.client.registrationapi.RegistrationApiClient {
+public class RegistrationApiClient extends RegistrationApiBindings {
     @Autowired
     private Environment env;
