+++ libraries/pom.xml
@@ -24,7 +24,7 @@
         <module>informationmodel-cat-bindings</module>
         <module>publisher-cat-client</module>
         <module>referencedata-client</module>
-        <module>registration-api-client</module>
+        <module>registration-api-bindings</module>
         <module>registration-api-common</module>
         <module>shared</module>
         <module>webutils-spring</module>
